% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pbapply}
\alias{pbapply}
\title{Apply with an auto-generated shiny progress bar.}
\usage{
pbapply(X, MARGIN, FUN, session = getDefaultReactiveDomain(),
  message = NULL, detail = NULL, ...)
}
\arguments{
\item{X}{An array, including a matrix that will be iterated
over.}

\item{MARGIN}{A vector giving the subscripts which the
function will be applied over. 1 indicates rows, 2 indicates
columns, c(1,2) indicates rows and columns.}

\item{FUN}{The function to apply to the elements of \code{X}.}

\item{session}{The Shiny session object, as provided by
shinyServer to the server function.}

\item{message}{A single-element character vector; the message
to be displayed to the user, or \code{NULL} to hide the current
message (if any).}

\item{detail}{A single-element character vector; the detail
message to be displayed to the user, or \code{NULL} to hide the
current detail message (if any). The detail message will be
shown with a de-emphasized appearance relative to \code{message}.}

\item{...}{Additional argument list that might not ever
 be used.}
}
\value{
A vector of results from \code{FUN} applied over \code{X}.
}
\description{
When called within a shiny app, this function will auto-
generate a progress bar, update the progress bar as evaluation
progresses, and close the progress bar upon completion.
}
\examples{
shiny::shinyApp( ui=fluidPage(titlePanel("Progress Bars!!"), actionButton('go', label="Start")),
                           server=shinyServer( function(input, output, session){
                                                      observe({
                                                          if(input$go > 0)
                                                             pbapply(matrix(c(1:20000000), ncol=20000), 1, FUN=summary, session=session, message="My Message")
                                                      })
                                                }) )
}
\seealso{
\code{\link{}} and \code{\link{}}
}

